+++ kernel/hash.hpp
@@ -43,8 +43,8 @@ namespace tuscany
  * tree at contrib/ltree/crc32.[ch] and from FreeBSD at
  * src/usr.bin/cksum/crc32.c.
  */
-const unsigned int crc32hash(const char* data, const unsigned int len) {
-    return apr_memcache_hash_default(NULL, data, len);
+const unsigned int crc32hash(const char* data, const size_t len) {
+    return (unsigned int)apr_memcache_hash_default(NULL, data, len);
 }
 
 /**
@@ -89,7 +89,7 @@ const unsigned int crc32hash(const char* data, const unsigned int len) {
  *
  *                  -- Ralf S. Engelschall <rse@engelschall.com>
  */
-const unsigned int times33hash(const char* data, const unsigned int len) {
+const unsigned int times33hash(const char* data, const size_t len) {
     apr_ssize_t l = len;
     return apr_hashfunc_default(data, &l);
 }
@@ -107,8 +107,8 @@ const unsigned int times33hash(const char* data, const unsigned int len) {
  * Passes Bob Jenkin's frog.c torture-test. No collisions possible for 4 byte
  * keys, no small 1 to 7 bit differentials.
  */
-const unsigned int murmurhash(const char* key, const unsigned int klen) {
-    unsigned int len = klen;
+const unsigned int murmurhash(const char* key, const size_t klen) {
+    unsigned int len = (unsigned int)klen;
     const unsigned int seed = 0;
 
     // 'm' and 'r' are mixing constants generated offline.
@@ -153,8 +153,8 @@ const unsigned int murmurhash(const char* key, const unsigned int klen) {
  * An endian and alignment neutral, but half the speed, version of
  * the murmur hash.
  */
-const unsigned int portablemurmurhash(const char* key, const unsigned int klen) {
-    unsigned int len = klen;
+const unsigned int portablemurmurhash(const char* key, const size_t klen) {
+    unsigned int len = (unsigned int)klen;
     const unsigned int seed = 0;
 
     // 'm' and 'r' are mixing constants generated offline.
